+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 * GGT-797 BOSS Properties removed and defaulted to SPECTRUM
 * GGT-797 Updated UAT TTL value to 90 days 
 * GGT-652 update gitlab hostname and protocol
+* GMB-817 Provision to add the lens content type for the files swept by Directory Sweeper
 
 4.4.0
 -------------------------------